+++ src/gen_rpc_client.erl
@@ -228,8 +228,8 @@ sbcast(Nodes, Name, Msg) when is_list(Nodes), is_atom(Name) ->
 init({Node}) ->
     ok = gen_rpc_helper:set_optimal_process_flags(),
     case gen_rpc_helper:get_client_config_per_node(Node) of
-        {error, {Class,Reason}} ->
-            ?log(error, "event=external_source_error action=falling_back_to_local reason=\"~s:~p\"", [Class, Reason]),
+        {error, Reason} ->
+            ?log(error, "event=external_source_error action=falling_back_to_local reason=\"~s\"", [Reason]),
             {stop, {badrpc, {external_source_error, Reason}}};
         {Driver, Port} ->
             {DriverMod, DriverClosed, DriverError} = gen_rpc_helper:get_client_driver_options(Driver),